Moved to 'install' instead of cp/chmod/chown

+++ install.sh
@@ -17,11 +17,11 @@ make all
 echo -en '\n'
 
 echo -en "${GREEN}* Sunwait installation"
-sudo cp sunwait /usr/local/bin
+sudo install sunwait /usr/local/bin/
 echo -en '\n'
 
 echo -en "${GREEN}* Using the camera without root access\n${NC}"
-sudo install asi.rules /etc/udev/rules.d
+sudo install -D -m 0644 -u root -g root asi.rules /etc/udev/rules.d/
 sudo udevadm control -R
 echo -en '\n'
 
@@ -29,18 +29,12 @@ echo -en "${GREEN}* Autostart script\n${NC}"
 sed -i '/allsky.sh/d' /etc/xdg/lxsession/LXDE-pi/autostart
 sed -i "s|User=pi|User=$USER|g" autostart/allsky.service
 sed -i "s|/home/pi/allsky|$PWD|g" autostart/allsky.service
-sudo cp autostart/allsky.service /etc/systemd/system/
-sudo chown root:root /etc/systemd/system/allsky.service
-sudo chmod 0644 /etc/systemd/system/allsky.service
+sudo install -D -m 0644 -o root -g root autostart/allsky.service /etc/systemd/system/
 echo -en '\n'
 
 echo -en "${GREEN}* Configure log rotation\n${NC}"
-sudo cp autostart/allsky /etc/logrotate.d/
-sudo chown root:root /etc/logrotate.d/allsky
-sudo chmod 0644 /etc/logrotate.d/allsky
-sudo cp autostart/allsky.conf /etc/rsyslog.d/ 
-sudo chown root:root /etc/rsyslog.d/allsky.conf
-sudo chmod 0644 /etc/rsyslog.d/allsky.conf
+sudo install -D -m 0644 -u root -g root autostart/allsky /etc/logrotate.d/
+sudo install -D -m 0644 -u root -g root autostart/allsky.conf /etc/rsyslog.d/ 
 echo -en '\n'
 
 echo -en "${GREEN}* Add ALLSKY_HOME environment variable\n${NC}"
@@ -48,10 +42,10 @@ echo "export ALLSKY_HOME=$PWD" | sudo tee /etc/profile.d/allsky.sh &> /dev/null
 echo -en '\n'
 
 echo -en "${GREEN}* Copy camera settings files\n${NC}"
-cp settings_ZWO.json.repo settings_ZWO.json
-cp settings_RPiHQ.json.repo settings_RPiHQ.json
-cp config.sh.repo config.sh
-cp scripts/ftp-settings.sh.repo scripts/ftp-settings.sh
+[ ! -e settings_ZWO.json ] && cp settings_ZWO.json.repo settings_ZWO.json
+[ ! -e settings_RPiHQ.json ] && cp settings_RPiHQ.json.repo settings_RPiHQ.json
+[ ! -e config.sh ] && cp config.sh.repo config.sh
+[ ! -e scripts/ftp-settings.sh ] && cp scripts/ftp-settings.sh.repo scripts/ftp-settings.sh
 sudo chown -R $USER:$USER ./
 sudo systemctl daemon-reload
 sudo systemctl enable allsky.service
